Created comps for each section of Order Details
